+++ frontend/src/scenes/insights/insightDataLogic.ts
@@ -38,7 +38,7 @@ import { insightVizDataNodeKey } from '~/queries/nodes/InsightViz/InsightViz'
 import { subscriptions } from 'kea-subscriptions'
 import { queryExportContext } from '~/queries/query'
 import { objectsEqual } from 'lib/utils'
-import { displayTypesWithoutLegend } from 'lib/components/InsightLegend/InsightLegend'
+import { displayTypesWithoutLegend } from 'lib/components/InsightLegend/utils'
 
 const defaultQuery = (insightProps: InsightLogicProps): Node => {
     const filters = insightProps.cachedInsight?.filters
